simplify the apl solution for day 7
the array of possible vertical positions that was generated was between
the minimum value to (maximum+minimum).  to make it the most optimal
solution, the array would have to be generated with:

 (⌊/i)+0,⍳(⌊/i)+⌈/i

but to make the solution a little shorter, and since the minimum value
in the test data and the input data is 0, the list is now just generated
with '0,⍳⌈/i', which produces an array from 0 to the maximum value.