simplify apl solution
the string literal 'abcdefghijklmnopqrstuvwxyz' has been replaced with
'(819⌶⎕a)'.  the variable 'vv' is also no longer defined, since it was
used only once inside the 'got' and the 'gottwo' dfns.

the 'got' and 'gottwo' dfns have also been put in a single line so that
the '12.apl' file can be run as a script.  because they are both used
only once, they are now also lambda functions.